+++ src/main/scala/org/openhorizon/exchangeapi/route/agreementbot/DeploymentPatterns.scala
@@ -52,7 +52,7 @@ trait DeploymentPatterns extends JacksonSupport with AuthenticationSupport {
                                        @Parameter(hidden = true) organization: String,
                                        @Parameter(hidden = true) resource: String): Route =
     delete {
-      logger.debug(s"DELETE /orgs/${organization}/agbots/${agreementBot}/patterns - By ${identity.resource}:${identity.role}")
+      logger.debug(s"DELETE /orgs/${organization}/agbots/${agreementBot}/patterns - ${identity.resource}:${identity.role}(${identity.identifier.getOrElse("")})(${identity.owner.getOrElse("")})")
       
       val INSTANT: Instant = Instant.now()
       
@@ -123,7 +123,7 @@ trait DeploymentPatterns extends JacksonSupport with AuthenticationSupport {
                                     @Parameter(hidden = true) identity: Identity2,
                                     @Parameter(hidden = true) organization: String,
                                     @Parameter(hidden = true) resource: String): Route = {
-    logger.debug(s"GET /orgs/${organization}/agbots/${agreementBot}/patterns - By ${identity.resource}:${identity.role}")
+    logger.debug(s"GET /orgs/${organization}/agbots/${agreementBot}/patterns - ${identity.resource}:${identity.role}(${identity.identifier.getOrElse("")})(${identity.owner.getOrElse("")})")
     complete({
       db.run(AgbotPatternsTQ.getPatterns(resource).result)
         .map({
@@ -173,7 +173,7 @@ trait DeploymentPatterns extends JacksonSupport with AuthenticationSupport {
     post {
       entity(as[PostAgbotPatternRequest]) {
         reqBody =>
-          logger.debug(s"POST /orgs/${organization}/agbots/${agreementBot}/patterns - By ${identity.resource}:${identity.role}")
+          logger.debug(s"POST /orgs/${organization}/agbots/${agreementBot}/patterns - ${identity.resource}:${identity.role}(${identity.identifier.getOrElse("")})(${identity.owner.getOrElse("")})")
             validateWithMsg(reqBody.getAnyProblem) {
               
               val INSTANT: Instant = Instant.now()