+++ src/main/scala/org/openhorizon/exchangeapi/route/agreementbot/DeploymentPolicy.scala
@@ -55,7 +55,7 @@ trait DeploymentPolicy extends JacksonSupport with AuthenticationSupport {
                                      @Parameter(hidden = true) identity: Identity2,
                                      @Parameter(hidden = true) organization: String,
                                      @Parameter(hidden = true) resource: String): Route = {
-    logger.debug(s"DELETE /orgs/${organization}/agbots/${agreementBot}/businesspols/${deploymentPolicy} - By ${identity.resource}:${identity.role}")
+    logger.debug(s"DELETE /orgs/${organization}/agbots/${agreementBot}/businesspols/${deploymentPolicy} - ${identity.resource}:${identity.role}(${identity.identifier.getOrElse("")})(${identity.owner.getOrElse("")})")
     
     val INSTANT: Instant = Instant.now()
     
@@ -121,7 +121,7 @@ trait DeploymentPolicy extends JacksonSupport with AuthenticationSupport {
                                   @Parameter(hidden = true) identity: Identity2,
                                   @Parameter(hidden = true) organization: String,
                                   @Parameter(hidden = true) resource: String): Route = {
-    logger.debug(s"GET /orgs/${organization}/agbots/${agreementBot}/businesspols/${deploymentPolicy} - By ${identity.resource}:${identity.role}")
+    logger.debug(s"GET /orgs/${organization}/agbots/${agreementBot}/businesspols/${deploymentPolicy} - ${identity.resource}:${identity.role}(${identity.identifier.getOrElse("")})(${identity.owner.getOrElse("")})")
     complete({
       db.run(AgbotBusinessPolsTQ.getBusinessPol(resource, deploymentPolicy).result)
         .map({