+++ src/main/scala/org/openhorizon/exchangeapi/route/search/NodeHealth.scala
@@ -107,7 +107,7 @@ trait NodeHealth extends JacksonSupport with AuthenticationSupport {
                            @Parameter(hidden = true) organization: String): Route =
     entity(as[PostNodeHealthRequest]) {
       reqBody =>
-        logger.debug(s"POST /orgs/$organization/search/nodehealth - By ${identity.resource}:${identity.role}")
+        logger.debug(s"POST /orgs/$organization/search/nodehealth - ${identity.resource}:${identity.role}(${identity.identifier.getOrElse("")})(${identity.owner.getOrElse("")})")
         validateWithMsg(reqBody.getAnyProblem) {
           complete({
             /*